backend/requirements.txt
attrs==19.3.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
decorator==4.4.1
dnslib==0.9.14
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==3.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.23.0
parso==0.6.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.3
pyrsistent==0.15.7
python-dotenv==0.15.0
pywin32==300
six==1.15.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
